Remove some useless controller and view to clean up a bit
