+++ app/Resources/views/base.html.twig
@@ -28,13 +28,8 @@
         <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300" rel="stylesheet">
 
         <!-- Favicon -->
-        <link rel="shortcut icon" type="image/x-icon" href="{{ asset('/je-donne-2-heures-favicon-ico.ico')}}" />
-        <link rel="icon" type="image/png" href="{{ asset('/je-donne-2-heures-favicon-16X16.png')}}" />
-
-
-
-        <link href="{{ asset('css/main.css')}}">
-
+        <link rel="shortcut icon" type="image/x-icon" href="{{ asset('img/je-donne-2-heures-favicon-ico.ico')}}"/>
+        <link rel="icon" type="image/png" href="{{ asset('img/je-donne-2-heures-favicon-16X16.png')}}" />
 
         <title>{% block title %}Welcome!{% endblock %}</title>
         {% block stylesheets %}{% endblock %}
