+++ app/Resources/views/donneur/index.html.twig
@@ -1,5 +1,7 @@
 {% extends 'back/baseback.html.twig' %}
 
+{%  block title %}Admin - Liste donneur | Je Donne 2 Heures{% endblock %}
+
 {% block body %}
     <h1>Liste des donneurs</h1>
