--- src/Jd2hBundle/Controller/DefaultController.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Jd2hBundle\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
-class DefaultController extends Controller
-{
-    public function indexAction()
-    {
-        return $this->render('Jd2hBundle:Default:index.html.twig');
-    }
-
-}
-
-
