+++ src/Jd2hBundle/Controller/FrontController.php
@@ -45,7 +45,7 @@ class FrontController extends Controller {
                 $message = \Swift_Message::newInstance();
                 $message->setSubject('Inscription à Je donne 2 heures')
                             ->setFrom(array('projet.jd2h@gmail.com' => 'Je donne 2 heures'))
-                            ->setTo($donneur->getMailAdress())
+                            ->setTo($donneur->getMailAddress())
                             ->setBody(
                             '<html>' .
                             '<head></head>' .
