+++ src/Jd2hBundle/Resources/config/routing/back.yml
@@ -1,20 +1,20 @@
-back_entrepreneur:
-    path:     /entrepreneur
-    defaults: { _controller: "Jd2hBundle:Back:entrepreneur" }
-    methods:  GET
+#back_entrepreneur:
+#    path:     /entrepreneur
+ #   defaults: { _controller: "Jd2hBundle:Back:entrepreneur" }
+  #  methods:  GET
 
-back_donneur:
-    path:     /donneur
-    defaults: { _controller: "Jd2hBundle:Back:donneur" }
-    methods:  GET
+#back_donneur:
+ #   path:     /donneur
+  #  defaults: { _controller: "Jd2hBundle:Back:donneur" }
+   # methods:  GET
 
 admin:
     path:     /admin
-    defaults: { _controller: "Jd2hBundle:Security:admin" }
+    defaults: { _controller: "Jd2hBundle:Back:admin" }
 
 login:
     path:     /login
-    defaults: { _controller: "Jd2hBundle:Security:login"}
+    defaults: { _controller: "Jd2hBundle:Back:login"}
 
 logout:
     path:     /logout
